🛠️ refactor: (BE) - room dto directory 위치 수정
