+++ backend/socket/src/room/dto/socket-response.dto.ts
